Ket 2

+++ src/main/java/com/example/DTO/ChiTietHoaDonDTO.java
@@ -6,6 +6,18 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+// @Data
+// @NoArgsConstructor
+// @AllArgsConstructor
+// public class ChiTietHoaDonDTO {
+//     private Integer id;
+//     private Date ngayDat;
+//     private String tenTour;
+//     private int soLuongNguoi;
+//     private double thanhTien; 
+//     private boolean trangThai;
+//     private boolean phuongThucThanhToan;
+// }
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,8 +25,8 @@ public class ChiTietHoaDonDTO {
     private Integer id;
     private Date ngayDat;
     private String tenTour;
-    private int soLuongNguoi;
-    private double thanhTien; 
+    private Long soLuongNguoi; // Thay đổi thành Long
+    private double thanhTien;
     private boolean trangThai;
     private boolean phuongThucThanhToan;
 }
